requirements.txt
google-generativeai>=0.8.0
selenium>=4.16.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
lxml>=5.1.0
